refactor(models/meta): dump meta types into mod.rs
they were split weirdly s.t. it was hard to find them lol
