+++ src/models/metadata/builder/mp4parse.rs
@@ -8,10 +8,7 @@ use sqlx::types::Json;
 
 use crate::{
     error::RavesError,
-    models::metadata::{
-        builder::get_video_len,
-        types::{Format, MediaKind},
-    },
+    models::metadata::{builder::get_video_len, Format, MediaKind},
 };
 
 use super::MediaBuilder;
